Agent changes and ex. Agent
Changed a lot around the agent-class, creating an example regarding the restaurant. Further developments and explanations will be embedded in the code, as well as functions regarding the class of the agent. So far, everything is implemented in the agent's class and the way it is structured, no second class is needed. This might still change, however right now I'm keeping it this way, as all sets of actions, possible states and connections between the two are unique to the agents anyway, shared actions can later be found out by a class method of the class agents, but this is still a bit fuzzy.