Implemented "act"-function into Agent
A function to act was implemented. It first computes what actions are possible based on the current state and in a second step calculates the utility based on the utility function, called "val_function". It also outputs information regarding all processes.
Next step will be to implement a change in the current state, so far not happening(so he'll be hungry forever and going to the michelin restaurant forever, making him the most obese king in Dutch history).