Added State change (+primary/secondary attributes)
I've implemented functionality to include status changes. We have to still talk about what implementation is the best, however in this case every action now has two different dictionarys attached to them. As previously there are "secondary attributes", like the price and quality etc., now there is a dictionary of "primary vattributes", basically referring to "things that change" in the state, in this case hungry obviously is changed to False. There are other ways to implement this though and there is also the possibility to have a 2-step-system of decision making here: First, the system decides on which of the actions satisfy as many of the primary targets, the state-changing targets, as possible, then based on the "secondary attributes" like price and quality, it chooses between the options that satisfy the same number of primary attributes(in thise case between all actions, because all satisfy the primary goal of erasing hunger).